Move ports to application layer
