+++ src/main/java/at/fhv/se/smartmeter/adapter/timescaledb/TimescaleOperationalPropertyDefRepository.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import at.fhv.se.smartmeter.adapter.timescaledb.jpaRepos.TimescaleJPAOperationalPropertyDef;
-import at.fhv.se.smartmeter.domain.port.outbound.persistence.OperationalPropertyDefRepository;
+import at.fhv.se.smartmeter.application.port.outbound.persistence.OperationalPropertyDefRepository;
 
 @Service
 public class TimescaleOperationalPropertyDefRepository implements OperationalPropertyDefRepository {
