+++ src/main/java/at/fhv/se/smartmeter/application/dto/MeterReadingQueryDTO.java
@@ -1,12 +1,12 @@
-package at.fhv.se.smartmeter.adapter.dto;
+package at.fhv.se.smartmeter.application.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
-
 @Getter
 @Builder
-public class MeterReadingDTO {
+public class MeterReadingQueryDTO {
+    private String id;
     private String readingTime;
     private String meterId;
     private MeterReadingPropDTO[] propertyValues;
