+++ src/main/java/at/fhv/se/smartmeter/application/port/inbound/meter/CreateMeterUseCase.java
@@ -1,4 +1,4 @@
-package at.fhv.se.smartmeter.domain.port.inbound.meter;
+package at.fhv.se.smartmeter.application.port.inbound.meter;
 
 public interface CreateMeterUseCase {
     String createMeter(String serialNumber, String physicalMeterId);
