+++ src/main/java/at/fhv/se/smartmeter/application/port/inbound/meterReading/CreateMeterReadingUseCase.java
@@ -1,0 +1,7 @@
+package at.fhv.se.smartmeter.application.port.inbound.meterReading;
+
+import at.fhv.se.smartmeter.application.dto.MeterReadingDTO;
+
+public interface CreateMeterReadingUseCase {
+    String createMeterReading(MeterReadingDTO meterReadingDTO);
+}
