+++ src/main/java/at/fhv/se/smartmeter/application/port/inbound/meterReading/GetMeterReadingForIntervalUseCase.java
@@ -1,8 +1,8 @@
-package at.fhv.se.smartmeter.domain.port.inbound.meterReading;
+package at.fhv.se.smartmeter.application.port.inbound.meterReading;
 
 import java.util.List;
 
-import at.fhv.se.smartmeter.adapter.dto.MeterReadingQueryDTO;
+import at.fhv.se.smartmeter.application.dto.MeterReadingQueryDTO;
 
 public interface GetMeterReadingForIntervalUseCase {
     List<MeterReadingQueryDTO> getMeterReadingsForInterval(String meterId, String startDate, String endDate);
