+++ src/main/java/at/fhv/se/smartmeter/application/port/outbound/persistence/MeterReadingRepository.java
@@ -1,4 +1,4 @@
-package at.fhv.se.smartmeter.domain.port.outbound.persistence;
+package at.fhv.se.smartmeter.application.port.outbound.persistence;
 
 import java.time.ZonedDateTime;
 import java.util.List;
