+++ src/main/java/at/fhv/se/smartmeter/application/port/outbound/persistence/OperationalPropertyDefRepository.java
@@ -1,4 +1,4 @@
-package at.fhv.se.smartmeter.domain.port.outbound.persistence;
+package at.fhv.se.smartmeter.application.port.outbound.persistence;
 
 public interface OperationalPropertyDefRepository {
     boolean existsById(String id);
