--- src/main/java/at/fhv/se/smartmeter/domain/port/inbound/meterReading/CreateMeterReadingUseCase.java
@@ -1,7 +1,0 @@
-package at.fhv.se.smartmeter.domain.port.inbound.meterReading;
-
-import at.fhv.se.smartmeter.adapter.dto.MeterReadingDTO;
-
-public interface CreateMeterReadingUseCase {
-    String createMeterReading(MeterReadingDTO meterReadingDTO);
-}
